Performance improvements to the individual profile's mappers and reducers.
git-svn-id: https://barricada.hi.inet/repositorio/ps2@148 72e40b64-262a-4186-abad-0444e9ae7036

+++ trunk/individualprofile/src/main/java/es/tid/ps/profile/userprofile/CategoryCount.java
@@ -17,7 +17,7 @@ public class CategoryCount implements Writable {
 
     public CategoryCount() {
         this.category = null;
-        this.count = 0;
+        this.count = 0l;
     }
 
     public CategoryCount(String category, long count) {
@@ -37,7 +37,7 @@ public class CategoryCount implements Writable {
         return this.count;
     }
 
-    public void setCount(int count) {
+    public void setCount(long count) {
         this.count = count;
     }
